+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * EAM-104 - Join and view buttons
 * KLS-388 - Added CTA streamfield block
 * KLS-467 - Update business details turnover options for the export contact page
+* KLS-269 - Develop frontend of events listing page
 
 
 ## [2.28.0](https://github.com/uktrade/great-cms/releases/tag/2.28.0)